requirements.txt
pypdf
requests
python-dotenv
stripe
